WIP: move some custom SA types into deprecated namespace
